refactor(query): decompose monolithic query builder
Closes THE-2440
